+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ### Features
 
+- Add `sentryzerolog` integration ([#857](https://github.com/getsentry/sentry-go/pull/857))
 - Add `sentryslog` integration ([#865](https://github.com/getsentry/sentry-go/pull/865))
-
 - Always set Mechanism Type to generic ([#896](https://github.com/getsentry/sentry-go/pull/897))
 
 ### Misc